requirements.txt
anyio==3.6.1; python_version >= "3.6" and python_full_version >= "3.6.2"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
fastapi==0.81.0; python_full_version >= "3.6.1"
h11==0.13.0; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.2"
importlib-metadata==4.12.0; python_version < "3.10" and python_version >= "3.7"
markdown==3.4.1; python_version >= "3.7"
numpy==1.23.2; python_version >= "3.8"
opencv-python==4.6.0.66; python_version >= "3.6"
packaging==21.3; python_version >= "3.7"
pillow==9.2.0; python_version >= "3.7"
pydantic==1.9.2; python_full_version >= "3.6.1"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7"
pytesseract==0.3.10; python_version >= "3.7"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.3.0; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "3.10"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
uvicorn==0.18.3; python_version >= "3.7"
zipp==3.8.1; python_version < "3.10" and python_version >= "3.7"
